requirements.txt..txt
fastapi
uvicorn
python-multipart
openai-whisper
transformers
torch
scipy